backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "scipy>=1.17.0",
    "selenium>=4.40.0",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "torch>=2.10.0",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
    "webdriver-manager>=4.0.2",
    "youtube-transcript-api>=1.2.3",
]
